+++ pyiron_workflow/create.py
@@ -67,7 +67,7 @@ class Creator(metaclass=Singleton):
     @property
     @lru_cache(maxsize=1)
     def for_node(self):
-        from pyiron_workflow.for_node import for_node
+        from pyiron_workflow.for_loop import for_node
         return for_node
 
     @property
